carve out version command in its own package
