+++ cmd/version/logo.txt
