+++ cmd/version/version_test.go
@@ -1,4 +1,4 @@
-package cmd
+package version
 
 import (
 	"testing"
